Use a file for all option types
