+++ demo/app.component.ts
@@ -195,7 +195,7 @@ export class AppComponent {
           {
             value: 'lost+found',
             options: {
-              icon: {
+              cssClasses: {
                 nodeExpanded: 'fa fa-circle',
                 nodeCollapsed: 'fa fa-circle',
                 nodeLeaf: 'fa fa-circle'
@@ -225,7 +225,7 @@ export class AppComponent {
             value: 'firstUser',
             options: {
               drag: false,
-              icon: {
+              cssClasses: {
                 nodeCollapsed: 'fa fa-folder',
                 nodeExpanded: 'fa fa-folder-open',
                 nodeLeaf: 'fa fa-file'
@@ -242,12 +242,12 @@ export class AppComponent {
                       {
                         value: 'bills',
                         children: [
-                          {value: '2016-07-01-mobile.pdf', options: {icon: {nodeLeaf: 'fa fa-file-pdf-o'}}},
-                          {value: '2016-07-01-electricity.pdf', options: {icon: {nodeLeaf: 'fa fa-file-excel-o'}}},
-                          {value: '2016-07-01-water.pdf', options: {icon: {nodeLeaf: 'fa fa-file-sound-o'}}},
-                          {value: '2016-07-01-internet.pdf', options: {icon: {nodeLeaf: 'fa fa-file-word-o'}}},
-                          {value: '2016-08-01-mobile.pdf', options: {icon: {nodeLeaf: 'fa fa-file-archive-o'}}},
-                          {value: '2016-10-01-internet.pdf', options: {icon: {nodeLeaf: 'fa fa-file-pdf-o'}}}
+                          {value: '2016-07-01-mobile.pdf', options: {cssClasses: {nodeLeaf: 'fa fa-file-pdf-o'}}},
+                          {value: '2016-07-01-electricity.pdf', options: {cssClasses: {nodeLeaf: 'fa fa-file-excel-o'}}},
+                          {value: '2016-07-01-water.pdf', options: {cssClasses: {nodeLeaf: 'fa fa-file-sound-o'}}},
+                          {value: '2016-07-01-internet.pdf', options: {cssClasses: {nodeLeaf: 'fa fa-file-word-o'}}},
+                          {value: '2016-08-01-mobile.pdf', options: {cssClasses: {nodeLeaf: 'fa fa-file-archive-o'}}},
+                          {value: '2016-10-01-internet.pdf', options: {cssClasses: {nodeLeaf: 'fa fa-file-pdf-o'}}}
                         ]
                       },
                       {value: 'photos', children: []}
@@ -265,7 +265,7 @@ export class AppComponent {
           {
             value: 'secondUser',
             options: {
-              icon: {
+              cssClasses: {
                 nodeCollapsed: 'fa fa-folder-o',
                 nodeExpanded: 'fa fa-folder-open-o',
                 nodeLeaf: 'fa fa-file-o'
@@ -307,12 +307,15 @@ export class AppComponent {
   };
 
   public filesOptions = {
-    icon: {
-      nodeCollapsed: 'fa fa-folder-o',
-      nodeExpanded: 'fa fa-folder-open-o',
-      nodeLeaf: 'fa fa-file-o'
+    cssClasses: {
+      nodeIcon: 'fa fa-folder-o',
+      leafIcon: 'fa fa-file-o',
+      nodeCollapsed: 'fa fa-arrow-right',
+      nodeExpanded: 'fa fa-arrow-down',
+      nodeLeaf: ''
     },
     rightMenu: false,
+    leftMenu: true,
     mainMenu: true,
     expanded: false,
     selectEvent: false,
