+++ src/tree.component.ts
@@ -1,5 +1,6 @@
-import { Input, Component, OnInit, EventEmitter, Output, ElementRef, Inject } from '@angular/core';
-import { TreeStatus, TreeModel, TreeModelOptions, TreeOptions, NodeIconOptions, FoldingType, NodeEvent, RenamableNode, NodeSelectedEvent } from './tree.types';
+import { Input, Component, OnInit, EventEmitter, Output, ViewChild, ElementRef, Inject } from '@angular/core';
+import { TreeStatus, TreeModel, FoldingType, NodeEvent, RenamableNode, NodeSelectedEvent } from './tree.types';
+import { TreeModelOptions, TreeOptions, MenuOptions } from './options.types';
 import { NodeDraggableService } from './draggable/node-draggable.service';
 import { NodeMenuService } from './menu/node-menu.service';
 import { NodeDraggableEventAction, NodeDraggableEvent } from './draggable/draggable.types';
